requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
plotly>=5.3.1
pandas>=1.3.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
sqlalchemy